perf(ebpf): kernel-offload direct flows per mode, cached per flow

+++ crates/honk-core/src/clash_api/ui.rs
@@ -436,8 +436,7 @@ fn parse_download_url(url: &str) -> anyhow::Result<(String, u16, String, bool)> 
         .ok_or_else(|| anyhow::anyhow!("empty host in external UI URL '{url}'"))?;
     let port = parsed
         .port_or_known_default()
-        .ok_or_else(|| anyhow::anyhow!("missing port in external UI URL '{url}'"))?
-        as u16;
+        .ok_or_else(|| anyhow::anyhow!("missing port in external UI URL '{url}'"))?;
 
     let mut path = parsed.path().to_string();
     if let Some(q) = parsed.query() {
